@@ -1,6 +1,8 @@
 module github.com/pjbgf/go-apparmor/example/code
 
-go 1.17
+go 1.18
+
+replace github.com/pjbgf/go-apparmor => ../..
 
 require (
 	github.com/bombsimon/logrusr/v2 v2.0.1
@@ -8,9 +10,7 @@ require (
 	github.com/sirupsen/logrus v1.8.1
 )
 
-replace github.com/pjbgf/go-apparmor => ../..
-
 require (
-	github.com/go-logr/logr v1.2.0 // indirect
-	golang.org/x/sys v0.0.0-20211107104306-e0b2ad06fe42 // indirect
+	github.com/go-logr/logr v1.2.3 // indirect
+	golang.org/x/sys v0.0.0-20220811171246-fbc7d0a398ab // indirect
 )
