@@ -1,8 +1,8 @@
 module github.com/pjbgf/go-apparmor
 
-go 1.17
+go 1.18
 
 require (
 	github.com/go-logr/logr v1.2.3
-	golang.org/x/sys v0.0.0-20220808155132-1c4a2a72c664
+	golang.org/x/sys v0.0.0-20220811171246-fbc7d0a398ab
 )
